feat!: remove legacy admin slash commands
Cuts every deprecated admin slash command. The collapsed `/config web`
launcher is the only admin slash command that remains; `WEBUI_ENABLED=true`
is now the sole path to administer the bot.

Removed slash commands and their handlers:
- /permissions, /setup, /announce, /announce-vc-stats, /poll,
  /reactrole, /notice, /dbtrunk, /vc, /botstats
- All /config subcommands except /config web (list, set, import,
  export, reset, reload)

Also removed:
- Wizard interaction handlers (src/handlers/wizard-*) — unreachable
  after /setup deletion
- Deprecation notice utility (src/utils/deprecation-notice.ts) and
  its hook in CommandManager.executeCommand — nothing left to flag
- Tests for all removed commands and handlers

Underlying services in src/services/ are intentionally kept; only the
slash registrations and command handlers are gone, leaving the WebUI
the only surface that drives them.

User-facing commands (/ping, /voicestats incl. top, /seen, /quote,
/achievements, /amikool, /help) continue to register and work.

Operators should run `npm run cleanup-global-commands` and/or
`npm run unregister-guild-commands` after deploying to drop stale
command registrations from Discord.

Closes #386

BREAKING CHANGE: All admin slash commands except `/config web` have been
removed. `WEBUI_ENABLED=true` (with the supporting env vars) is now
required to administer the bot. Operators must enable the WebUI before
upgrading to this release.
