--- __tests__/commands/permissions.test.ts
@@ -1,254 +1,0 @@
-import { describe, it, expect, jest } from "@jest/globals";
-import { data, autocomplete } from "../../src/commands/permissions.js";
-
-jest.mock('../../src/utils/logger.js');
-
-describe("Permissions Command", () => {
-  describe("command metadata", () => {
-    it("should have correct command name", () => {
-      expect(data.name).toBe("permissions");
-    });
-
-    it("should have a description", () => {
-      expect(data.description).toBe("Manage command permissions");
-    });
-
-    it("should require Administrator permission", () => {
-      const json = data.toJSON();
-      expect(json.default_member_permissions).toBeDefined();
-    });
-
-    it("should be a valid slash command with subcommands", () => {
-      const json = data.toJSON();
-      expect(json).toHaveProperty("name", "permissions");
-      expect(json).toHaveProperty("description", "Manage command permissions");
-      expect(json).toHaveProperty("options");
-      expect(Array.isArray(json.options)).toBe(true);
-    });
-  });
-
-  describe("subcommands", () => {
-    it("should have set subcommand", () => {
-      const json = data.toJSON();
-      const setSubcommand = json.options?.find(
-        (opt: { name: string }) => opt.name === "set",
-      );
-      expect(setSubcommand).toBeDefined();
-      expect(setSubcommand?.description).toBe(
-        "Set command permissions (replaces existing)",
-      );
-    });
-
-    it("should have add subcommand", () => {
-      const json = data.toJSON();
-      const addSubcommand = json.options?.find(
-        (opt: { name: string }) => opt.name === "add",
-      );
-      expect(addSubcommand).toBeDefined();
-      expect(addSubcommand?.description).toBe(
-        "Add roles to existing command permissions",
-      );
-    });
-
-    it("should have remove subcommand", () => {
-      const json = data.toJSON();
-      const removeSubcommand = json.options?.find(
-        (opt: { name: string }) => opt.name === "remove",
-      );
-      expect(removeSubcommand).toBeDefined();
-      expect(removeSubcommand?.description).toBe(
-        "Remove roles from command permissions",
-      );
-    });
-
-    it("should have clear subcommand with optional parameters", () => {
-      const json = data.toJSON();
-      const clearSubcommand = json.options?.find(
-        (opt: { name: string }) => opt.name === "clear",
-      );
-      expect(clearSubcommand).toBeDefined();
-      expect(clearSubcommand?.description).toContain("Clear permissions");
-      
-      // Check that command, role, and user options exist and are optional
-      const commandOption = clearSubcommand?.options?.find(
-        (opt: { name: string }) => opt.name === "command",
-      );
-      expect(commandOption).toBeDefined();
-      expect(commandOption?.required).toBe(false);
-
-      const roleOption = clearSubcommand?.options?.find(
-        (opt: { name: string }) => opt.name === "role",
-      );
-      expect(roleOption).toBeDefined();
-      expect(roleOption?.required).toBe(false);
-
-      const userOption = clearSubcommand?.options?.find(
-        (opt: { name: string }) => opt.name === "user",
-      );
-      expect(userOption).toBeDefined();
-      expect(userOption?.required).toBe(false);
-    });
-
-    it("should have list subcommand with optional command filter", () => {
-      const json = data.toJSON();
-      const listSubcommand = json.options?.find(
-        (opt: { name: string }) => opt.name === "list",
-      );
-      expect(listSubcommand).toBeDefined();
-      expect(listSubcommand?.description).toBe("List command permissions");
-      
-      // Check that command option exists and is optional
-      const commandOption = listSubcommand?.options?.find(
-        (opt: { name: string }) => opt.name === "command",
-      );
-      expect(commandOption).toBeDefined();
-      expect(commandOption?.required).toBe(false);
-    });
-
-    it("should have view subcommand", () => {
-      const json = data.toJSON();
-      const viewSubcommand = json.options?.find(
-        (opt: { name: string }) => opt.name === "view",
-      );
-      expect(viewSubcommand).toBeDefined();
-      expect(viewSubcommand?.description).toBe(
-        "View permissions for a user or role",
-      );
-    });
-  });
-
-  describe("set subcommand options", () => {
-    it("should support multiple role options", () => {
-      const json = data.toJSON();
-      const setSubcommand = json.options?.find(
-        (opt: { name: string }) => opt.name === "set",
-      );
-
-      const roleOptions = setSubcommand?.options?.filter(
-        (opt: { name: string }) => opt.name.startsWith("role"),
-      );
-
-      // Should have 5 role options (role1-role5)
-      expect(roleOptions?.length).toBe(5);
-
-      // First role should be required
-      const role1 = setSubcommand?.options?.find(
-        (opt: { name: string }) => opt.name === "role1",
-      );
-      expect(role1?.required).toBe(true);
-
-      // Other roles should be optional
-      const role2 = setSubcommand?.options?.find(
-        (opt: { name: string }) => opt.name === "role2",
-      );
-      expect(role2?.required).toBe(false);
-    });
-  });
-
-  describe("autocomplete", () => {
-    it("should export autocomplete function", () => {
-      expect(autocomplete).toBeDefined();
-      expect(typeof autocomplete).toBe("function");
-    });
-
-    it("should filter commands based on user input", async () => {
-      const mockCommands = new Map([
-        ["quote", {}],
-        ["vcstats", {}],
-        ["permissions", {}],
-        ["ping", {}],
-      ]);
-
-      const mockInteraction = {
-        client: {
-          commands: mockCommands,
-        },
-        options: {
-          getFocused: jest.fn().mockReturnValue({
-            name: "command",
-            value: "qu",
-          }),
-        },
-        respond: jest.fn(),
-      };
-
-      await autocomplete(mockInteraction as unknown as import("discord.js").AutocompleteInteraction);
-
-      expect(mockInteraction.respond).toHaveBeenCalledWith(
-        expect.arrayContaining([
-          expect.objectContaining({ name: "quote", value: "quote" }),
-        ]),
-      );
-      expect(mockInteraction.respond).not.toHaveBeenCalledWith(
-        expect.arrayContaining([
-          expect.objectContaining({ name: "ping", value: "ping" }),
-        ]),
-      );
-    });
-
-    it("should limit results to 25 items", async () => {
-      const mockCommands = new Map();
-      for (let i = 0; i < 30; i++) {
-        mockCommands.set(`command${i}`, {});
-      }
-
-      const mockInteraction = {
-        client: {
-          commands: mockCommands,
-        },
-        options: {
-          getFocused: jest.fn().mockReturnValue({
-            name: "command",
-            value: "",
-          }),
-        },
-        respond: jest.fn(),
-      };
-
-      await autocomplete(mockInteraction as unknown as import("discord.js").AutocompleteInteraction);
-
-      const respondCall = mockInteraction.respond.mock.calls[0][0];
-      expect(Array.isArray(respondCall)).toBe(true);
-      expect(respondCall.length).toBeLessThanOrEqual(25);
-    });
-
-    it("should handle errors gracefully", async () => {
-      const mockInteraction = {
-        client: {
-          commands: new Map([["quote", {}]]),
-        },
-        options: {
-          getFocused: jest.fn().mockImplementation(() => {
-            throw new Error("Test error");
-          }),
-        },
-        respond: jest.fn(),
-      };
-
-      await expect(
-        autocomplete(mockInteraction as unknown as import("discord.js").AutocompleteInteraction),
-      ).resolves.not.toThrow();
-    });
-
-    it("should only respond to command option autocomplete", async () => {
-      const mockInteraction = {
-        client: {
-          commands: new Map([["quote", {}]]),
-        },
-        options: {
-          getFocused: jest.fn().mockReturnValue({
-            name: "other",
-            value: "test",
-          }),
-        },
-        respond: jest.fn(),
-      };
-
-      await autocomplete(mockInteraction as unknown as import("discord.js").AutocompleteInteraction);
-
-      expect(mockInteraction.respond).not.toHaveBeenCalled();
-    });
-  });
-
-  // Execute tests removed - service mocking issues with getInstance().mockReturnValue
-});
