--- __tests__/commands/setup-wizard-helpers.test.ts
@@ -1,11 +1,0 @@
-import { describe, it, expect } from '@jest/globals';
-
-describe('Setup Wizard Helpers', () => {
-  describe('module structure', () => {
-    it('placeholder test', () => {
-      // Module import causes hang due to WizardService.getInstance() call
-      // Simplified to placeholder test to avoid circular dependency
-      expect(true).toBe(true);
-    });
-  });
-});
