--- __tests__/commands/setup-wizard-integration.test.ts
@@ -1,70 +1,0 @@
-import { describe, it, expect } from '@jest/globals';
-
-/**
- * Integration test for setup wizard quotes configuration
- * This test validates the fix for the issue where running `/setup wizard quotes`
- * would show "Feature configuration for quotes is being set up..." but then nothing happened.
- * 
- * Note: Direct module imports are avoided due to singleton service initialization
- * (WizardService.getInstance(), ConfigService.getInstance()) at module level which
- * causes test environment issues. This is consistent with other wizard tests in the codebase.
- */
-describe('Setup Wizard Integration', () => {
-  describe('module structure', () => {
-    it('should validate fix was applied', () => {
-      // The fix involved:
-      // 1. Moving configuration functions from setup-wizard.ts to setup-wizard-helpers.ts
-      // 2. Implementing complete startFeatureConfiguration with switch statement
-      // 3. Exporting FEATURES constant to eliminate duplication
-      // 4. Adding proper type safety (Guild, ChatInputCommandInteraction)
-      
-      // This test documents the fix structure without importing modules that
-      // would fail due to service singleton initialization at module level.
-      expect(true).toBe(true);
-    });
-
-    it('should document new features added', () => {
-      // The following features were added to the wizard:
-      const newFeatures = ['amikool', 'reactionroles', 'announcements'];
-      
-      // Each feature has:
-      // - name: Human readable name
-      // - emoji: Visual identifier
-      // - description: Feature description
-      // - configKey: Config property to check enabled status (e.g., "amikool.enabled")
-      
-      expect(newFeatures.length).toBe(3);
-      expect(newFeatures).toContain('amikool');
-      expect(newFeatures).toContain('reactionroles');
-      expect(newFeatures).toContain('announcements');
-    });
-
-    it('should document feature status indicators', () => {
-      // Status indicators were added:
-      // ✅ = Feature is enabled
-      // ⚪ = Feature is disabled
-      // 
-      // These are shown in:
-      // 1. Embed field names
-      // 2. Select menu option labels
-      
-      expect(true).toBe(true);
-    });
-
-    it('should document channel selection pagination', () => {
-      // Pagination was added to handle Discord's 25-option limit in select menus:
-      // - WizardState now includes channelPage field for tracking current page
-      // - showChannelSelectionPage helper function displays paginated channels
-      // - Previous/Next buttons allow navigation through channel pages
-      // - Supports pagination for quotes, logging, and voice category selection
-      // - Displays "Showing X-Y of Z" and "Page N of M" in embed description
-      
-      const MAX_OPTIONS_PER_PAGE = 25; // Discord's select menu limit
-      expect(MAX_OPTIONS_PER_PAGE).toBe(25);
-      
-      // Pagination is applied to these selection types:
-      const paginatedSelections = ['quotes', 'logging', 'vc_category'];
-      expect(paginatedSelections.length).toBe(3);
-    });
-  });
-});
