--- __tests__/commands/vc.test.ts
@@ -1,73 +1,0 @@
-import { describe, it, expect } from "@jest/globals";
-import { data } from "../../src/commands/vc.js";
-
-describe("VC Command", () => {
-  describe("command metadata", () => {
-    it("should have correct command name", () => {
-      expect(data.name).toBe("vc");
-    });
-
-    it("should have a description", () => {
-      expect(data.description).toBe("Voice channel management");
-    });
-
-    it("should require administrator permissions", () => {
-      const json = data.toJSON();
-      expect(json).toHaveProperty("default_member_permissions");
-      // Administrator permission bit is "8"
-      expect(json.default_member_permissions).toBe("8");
-    });
-
-    it("should be a valid slash command", () => {
-      const json = data.toJSON();
-      expect(json).toHaveProperty("name", "vc");
-      expect(json).toHaveProperty("description");
-    });
-  });
-
-  describe("subcommands", () => {
-    const json = data.toJSON();
-    const options = json.options || [];
-
-    it("should have reload subcommand", () => {
-      const reloadSubcommand = options.find(
-        (opt: any) => opt.type === 1 && opt.name === "reload",
-      );
-      expect(reloadSubcommand).toBeDefined();
-      expect(reloadSubcommand?.description).toBe(
-        "Clean up empty voice channels",
-      );
-    });
-
-    it("should have force-reload subcommand", () => {
-      const forceReloadSubcommand = options.find(
-        (opt: any) => opt.type === 1 && opt.name === "force-reload",
-      );
-      expect(forceReloadSubcommand).toBeDefined();
-      expect(forceReloadSubcommand?.description).toBe(
-        "Force cleanup of ALL unmanaged channels in category",
-      );
-    });
-
-    it("should have exactly 2 admin subcommands", () => {
-      expect(options).toHaveLength(2);
-      const subcommandNames = options.map((opt: any) => opt.name);
-      expect(subcommandNames).toEqual(
-        expect.arrayContaining(["reload", "force-reload"]),
-      );
-    });
-
-    it("should not have customize subcommand group", () => {
-      // Verify customize group was removed (moved to control panel)
-      const customizeGroup = options.find(
-        (opt: any) => opt.type === 2 && opt.name === "customize",
-      );
-      expect(customizeGroup).toBeUndefined();
-    });
-  });
-
-  describe('execute', () => {
-    // Tests removed due to service initialization/mocking challenges
-    // Focus on command structure validation instead
-  });
-});
